// app/src/api/jupiter-v4/api.ts
export interface Def0LpFee {
    /**
     *
     * @type {string}
     * @memberof Def0LpFee
     */
    amount?: string;
    /**
     *
     * @type {string}
     * @memberof Def0LpFee
     */
    mint?: string;
    /**
     *
     * @type {number}
     * @memberof Def0LpFee
     */
    pct?: number;
}
/**
 *
 * @export
 * @interface Def1
 */
export interface Def1 {
    /**
     *
     * @type {string}
     * @memberof Def1
     */
    inAmount: string;
    /**
     *
     * @type {string}
     * @memberof Def1
     */
    outAmount: string;
    /**
     *
     * @type {number}
     * @memberof Def1
     */
    priceImpactPct: number;
    /**
     *
     * @type {Array<Def1MarketInfos>}
     * @memberof Def1
     */
    marketInfos: Array<Def1MarketInfos>;
    /**
     *
     * @type {string}
     * @memberof Def1
     */
    amount: string;
    /**
     *
     * @type {number}
     * @memberof Def1
     */
    slippageBps: number;
    /**
     * The threshold for the swap based on the provided slippage: when swapMode is ExactIn the minimum out amount, when swapMode is ExactOut the maximum in amount
     * @type {string}
     * @memberof Def1
     */
    otherAmountThreshold: string;
    /**
     *
     * @type {string}
     * @memberof Def1
     */
    swapMode: Def1.SwapModeEnum;
    /**
     *
     * @type {Def1Fees}
     * @memberof Def1
     */
    fees?: Def1Fees;
}

/**
 * @export
 * @namespace Def1
 */
export namespace Def1 {
    /**
     * @export
     * @enum {string}
     */
    export enum SwapModeEnum {
        ExactIn = <any> 'ExactIn',
        ExactOut = <any> 'ExactOut'
    }
}
/**
 * Only returned when userPublicKey is given to /quote
 * @export
 * @interface Def1Fees
 */
export interface Def1Fees {
    /**
     * This inidicate the total amount needed for signing transaction(s). Value in lamports.
     * @type {number}
     * @memberof Def1Fees
     */
    signatureFee?: number;
    /**
     * This inidicate the total amount needed for deposit of serum order account(s). Value in lamports.
     * @type {Array<number>}
     * @memberof Def1Fees
     */
    openOrdersDeposits?: Array<number>;
    /**
     * This inidicate the total amount needed for deposit of associative token account(s). Value in lamports.
     * @type {Array<number>}
     * @memberof Def1Fees
     */
    ataDeposits?: Array<number>;
    /**
     * This inidicate the total lamports needed for fees and deposits above.
     * @type {number}
     * @memberof Def1Fees
     */
    totalFeeAndDeposits?: number;
    /**
     * This inidicate the minimum lamports needed for transaction(s). Might be used to create wrapped SOL and will be returned when the wrapped SOL is closed.
     * @type {number}
     * @memberof Def1Fees
     */
    minimumSOLForTransaction?: number;
}
/**
 *
 * @export
 * @interface Def1MarketInfos
 */
export interface Def1MarketInfos {
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    id?: string;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    label?: string;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    inputMint?: string;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    outputMint?: string;
    /**
     *
     * @type {boolean}
     * @memberof Def1MarketInfos
     */
    notEnoughLiquidity?: boolean;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    inAmount?: string;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    outAmount?: string;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    minInAmount?: string;
    /**
     *
     * @type {string}
     * @memberof Def1MarketInfos
     */
    minOutAmount?: string;
    /**
     *
     * @type {number}
     * @memberof Def1MarketInfos
     */
    priceImpactPct?: number;
    /**
     *
     * @type {Def0LpFee}
     * @memberof Def1MarketInfos
     */
    lpFee?: Def0LpFee;
    /**
     *
     * @type {Def0LpFee}
     * @memberof Def1MarketInfos
     */
    platformFee?: Def0LpFee;
}